Move Console module to own file, rename files with name of class they contain
